+++ changelog.md
@@ -4,6 +4,10 @@
 ### Additions
 * Copper tools now show depth bonuses
 
+### Changes
+* Veining now also affects right-click actions
+  * Strip, pathing, tilling, etc
+
 ### Fixes
 * Copper depth bonus being completely broken
 * Fixed Veining breaking one block less than excpected